Feather-ai/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.19
django-allauth==0.54.0
django-otp==1.2.0
django-otp-twilio==1.0.3
django-verification==1.2.0
djangorestframework==3.14.0
google-api-core==2.11.1
google-api-python-client==2.48.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfiglet==0.8.post1
PyJWT==2.7.0
pyotp==2.8.0
pyparsing==3.0.9
python-resize-image==1.1.20
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==1.1.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wikipedia==1.4.0
windows-curses==2.3.1
